requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.1
python-dotenv>=1.0.1
streamlit-ace>=0.1.1
